Removing Account bug fixed due to Reference accounts ...

+++ lib/Model/Account.php
@@ -113,6 +113,7 @@ class Model_Account extends Model_Table {
 		$this->hasMany('AccountGuarantor','account_id');
 		$this->hasMany('TransactionRow','account_id');
 		$this->hasMany('Account','related_account_id',null,'RelatedAccounts');
+		$this->hasMany('Transaction','reference_id',null,'RelatedTransactions');
 		$this->hasMany('Comment','account_id');
 
 		$this->addHook('beforeSave',array($this,'defaultBeforeSave'));
@@ -163,12 +164,23 @@ class Model_Account extends Model_Table {
 		if($this->table !='accounts_pending' AND $this->ref('TransactionRow')->count()->getOne() > 0)
 			throw $this->exception('Account Contains Transactions, Cannot Delete');
 
+		if($related_tran = $this->ref('RelatedTransactions')->addCondition('cr_sum','<>',0)->count()->getOne()){
+			$related_tran = $this->ref('RelatedTransactions')->addCondition('cr_sum','<>',0)->tryLoadAny();
+			throw $this->exception('Related Transaction found')
+						->addMoreInfo('Narartion',$related_tran['Narration'])
+						->addMoreInfo('created_at',$related_tran['created_at'])
+						->addMoreInfo('In Branch',$related_tran['branch'])
+						->addMoreInfo('Transaction Type',$related_tran['transaction_type']);
+		}
+
+
 		$this->ref('Premium')->deleteAll();
 		$this->ref('JointMember')->deleteAll();
 		// $this->ref('Premium')->deleteAll();
 		$this->ref('DocumentSubmitted')->deleteAll();
 		$this->ref('AccountGuarantor')->deleteAll();
 		$this->ref('Comment')->deleteAll();
+		$this->ref('RelatedTransactions')->deleteAll();
 
 	}
 
@@ -835,7 +847,7 @@ class Model_Account extends Model_Table {
 		$transactions->addCondition('account_id',$this->id);
 
 		foreach ($transactions as $transactions_array) {
-			$transactions->delete(true);
+			$transactions->foreceDelete();
 		}
 	}
